% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{admin1.coordinates}
\alias{admin1.coordinates}
\title{World coordinates at the Admin-1 level}
\format{A SpatialPolygonsDataFrame with data frame with 4647 rows and 59 columns:}
\source{
\url{http://www.naturalearthdata.com/downloads/10m-cultural-vectors/}
}
\usage{
admin1.coordinates
}
\description{
World coordinates at the Admin-1 level
}
\keyword{datasets}

