% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/scatterplot.R
\name{BoxCenter}
\alias{BoxCenter}
\title{Box center
\code{BoxCenter} Computes the coordinates of the center of a box based upon the coordinates of the corner.}
\usage{
BoxCenter(ax, ay, bx, by, cx, cy, dx, dy)
}
\arguments{
\item{ax}{The x coordinate of the bottom-left of the box.}

\item{ay}{The y coordinate of the bottom-left of the box.}

\item{bx}{The x coordinate of the bottom-right of the box.}

\item{by}{The y coordinate of the bottom-right of the box.}

\item{cx}{The x coordinate of the top-right of the box.}

\item{cy}{The y coordinate of the top-right of the box.}

\item{dx}{The x coordinate of the top-left of the box.}

\item{dy}{The y coordinate of the top-left of the box.}
}
\value{
vector containing the x,y coordinates of the center.
}
\description{
Box center
\code{BoxCenter} Computes the coordinates of the center of a box based upon the coordinates of the corner.
}

