% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/labeledscatterplot.R
\name{BoxCoordinates}
\alias{BoxCoordinates}
\title{#' \code{rectangleOverlap} Checks to see if two recangles.
#'
#' @param a The first rectangle.
#' @param b The second recangle.
rectangleOverlap <- function(a, b)
{
    a$x < b$x + b$width &
        a$x + a$width > b$x &
        a$y < b$y + b$height &
        a$y + a$height > b$y;
}}
\usage{
BoxCoordinates(x, y, w, h)
}
\arguments{
\item{x}{x coordinate of the middle of the box.}

\item{y}{y coordinate of the middle of the box.}

\item{w}{width of the box in the x-scale.}

\item{h}{height  of the box in the y-scale.}
}
\value{
A vector containing the coordinates of a box.
}
\description{
#' \code{rectangle} The coordinates of a rectangle: bottom-left corner, width and height.
#'
#' @param x x coordinate of the middle of the box.
#' @param y y coordinate of the middle of the box.
#' @param w width of the box in the x-scale.
#' @param h height  of the box in the y-scale.
#' @return A list containing the coordinates of a box, consisting
#' of the bottom left coordinates, width and height.
rectangle = function(x, y, w, h)
{
  result <- list(x = x - w / 2, y = y - h / 2, width = w, height = h)
  class(result) <- "rectangle"
  result
}
}
\details{
Box coordinates
\code{BoxOverlap} Computes the positions of the corner of a box.
}

