% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/scatterplot.R
\name{BoxOverlap}
\alias{BoxOverlap}
\title{Whether or not two boxes overlap
\code{BoxOverlap} Checks to see if two boxes overlap}
\usage{
BoxOverlap(ax1, ay1, bx1, by1, cx1, cy1, dx1, dy1, ax2, ay2, bx2, by2, cx2, cy2,
  dx2, dy2)
}
\arguments{
\item{ax1}{The x coordinate of the bottom-left of the first box.}

\item{ay1}{The y coordinate of the bottom-left of the first box.}

\item{bx1}{The x coordinate of the bottom-right of the first box.}

\item{by1}{The y coordinate of the bottom-right of the first box.}

\item{cx1}{The x coordinate of the top-right of the first box.}

\item{cy1}{The y coordinate of the top-right of the first box.}

\item{dx1}{The x coordinate of the top-left of the first box.}

\item{dy1}{The y coordinate of the top-left of the first box.}

\item{ax2}{The x coordinate of the bottom-left of the second box.}

\item{ay2}{The y coordinate of the bottom-left of the second box.}

\item{bx2}{The x coordinate of the bottom-right of the second box.}

\item{by2}{The y coordinate of the bottom-right of the second box.}

\item{cx2}{The x coordinate of the top-right of the second box.}

\item{cy2}{The y coordinate of the top-right of the second box.}

\item{dx2}{The x coordinate of the top-left of the second box.}

\item{dy2}{The y coordinate of the top-left of the second box.}
}
\value{
true if the boxes overlap and false otherwise.
}
\description{
Whether or not two boxes overlap
\code{BoxOverlap} Checks to see if two boxes overlap
}

