% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sankeydiagram.R
\name{categorizeVariable}
\alias{categorizeVariable}
\title{categorizeVariable}
\usage{
categorizeVariable(x, max.categories, var.name)
}
\arguments{
\item{x}{A variable.}

\item{max.categories}{When the number of unique values}

\item{var.name}{The name of the variable.
of numeric data exceeds this value, the variable is quantized.}
}
\description{
Quantizes a numeric variable.
}

