% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{country.coordinates}
\alias{country.coordinates}
\title{Prices of 50,000 round cut diamonds.}
\format{A SpatialPolygonsDataFrame with data frame with 177 rows and 63 columns:}
\source{
\url{#https://raw.github.com/datasets/geo-boundaries-world-110m/master/countries.geojson}
}
\usage{
country.coordinates
}
\description{
A dataset containing the prices and other attributes of almost 54,000
diamonds.
}
\keyword{datasets}

