% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/interactivelabeledscatterplot.R
\name{createCaTooltips}
\alias{createCaTooltips}
\title{\code{createTooltips}}
\usage{
createCaTooltips(x, rowlabels, columnlabels)
}
\arguments{
\item{x}{The data that is being analzyed.}

\item{row.labels}{A vector of the row labels.}

\item{column.labels}{A vector of the column labels.}
}
\description{
Creates tooltips for \code{\link{CorrespondenceAnalysis}}.
}

