% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/interactivelabeledscatterplot.R
\name{CreateInteractiveScatterplotTooltips}
\alias{CreateInteractiveScatterplotTooltips}
\title{\code{CreateInteractiveScatterplotTooltips}}
\usage{
CreateInteractiveScatterplotTooltips(x)
}
\arguments{
\item{x}{The data that is being analzyed, where each row corresponds to a tooltip and
and the cells in that row appear in the tooltip.}

\item{row.labels}{A vector of the row labels.}
}
\description{
Creates tooltips for \code{\link{InteractiveLabeledScatterPlot}}.
}

