% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{admin1.coordinates}
\alias{admin1.coordinates}
\alias{brand.associations}
\alias{cola.perceptions}
\alias{country.coordinates}
\alias{csd.perceptions}
\alias{q.colors}
\title{#' Admin-0 level coordinates for the world
#'
#' @format A SpatialPolygonsDataFrame with data frame with 177 rows and 63 columns:
#' @source \url{http://www.naturalearthdata.com/downloads/50m-cultural-vectors/}
#' @rdname data
"map.coordinates"
Admin-1 level coordinates for the world}
\format{A SpatialPolygonsDataFrame with data frame with 4647 rows and 59 columns:}
\source{
\url{http://www.naturalearthdata.com/downloads/10m-cultural-vectors/}

\url{http://www.naturalearthdata.com/downloads/10m-cultural-vectors/}

Numbers International Pty Ltd.

Numbers International Pty Ltd.

Numbers International Pty Ltd.

Numbers International Pty Ltd.
}
\usage{
admin1.coordinates

country.coordinates

brand.associations

cola.perceptions

csd.perceptions

q.colors
}
\description{
#' Admin-0 level coordinates for the world
#'
#' @format A SpatialPolygonsDataFrame with data frame with 177 rows and 63 columns:
#' @source \url{http://www.naturalearthdata.com/downloads/50m-cultural-vectors/}
#' @rdname data
"map.coordinates"
Admin-1 level coordinates for the world

Admin-1 level coordinates for the world

Brand Associations

Brand Associations

CSD Associations

Default Q chart types.
}
\keyword{datasets}

