% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/labeledscatterplot.R
\name{Distance}
\alias{Distance}
\title{Distance between points
\code{Distance} A Euclidean distance between points}
\usage{
Distance(e1x, e1y, e2x, e2y)
}
\arguments{
\item{e1x}{The x coordinate of a point.}

\item{e1y}{The y coordinate of a point.}

\item{e2x}{The x coordinate of another point.}

\item{e2y}{The y coordinate of another point.}
}
\value{
The distance.
}
\description{
Distance between points
\code{Distance} A Euclidean distance between points
}

