% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/scatterplot.R
\name{DistanceToBox}
\alias{DistanceToBox}
\title{Distance to a line segment
\code{DistanceToBox} Computes the distance between a point and the nearest point on a box (0 if in the box)}
\usage{
DistanceToBox(px, py, ax, ay, bx, by, cx, cy, dx, dy)
}
\arguments{
\item{px}{The x coordinate of a point.}

\item{py}{The y coordinate of a point.}

\item{ax}{The x coordinate of the bottom-left of the box.}

\item{ay}{The y coordinate of the bottom-left of the box.}

\item{bx}{The x coordinate of the bottom-right of the box.}

\item{by}{The y coordinate of the bottom-right of the box.}

\item{cx}{The x coordinate of the top-right of the box.}

\item{cy}{The y coordinate of the top-right of the box.}

\item{dx}{The xcoordinates of the top-left of the box.}

\item{dy}{The y coordinate of the top-left of the box.}
}
\value{
The squared distance.
}
\description{
Distance to a line segment
\code{DistanceToBox} Computes the distance between a point and the nearest point on a box (0 if in the box)
}

