% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/scatterplot.R
\name{DistanceToLineSegment}
\alias{DistanceToLineSegment}
\title{Distance to a line segment
\code{DistanceToLineSegment} Computes the distance between a point and the nearest point on a line segment}
\usage{
DistanceToLineSegment(px, py, ax, ay, bx, by)
}
\arguments{
\item{px}{The x coordinate of a point.}

\item{py}{The y coordinate of a point.}

\item{ax}{The x coordinate of one end of a line.}

\item{ay}{The y coordinate of one end of a line.}

\item{bx}{The x coordinate of the other end of a line.}

\item{by}{The y coordinate of the other end of a line.}
}
\value{
The distance.
}
\description{
Distance to a line segment
\code{DistanceToLineSegment} Computes the distance between a point and the nearest point on a line segment
}

