% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sankeydiagram.R
\name{findNodesToMerge}
\alias{findNodesToMerge}
\title{findNodesToMerge}
\usage{
findNodesToMerge(df, column, weights = NULL)
}
\arguments{
\item{df}{dataframe containing factor variables only.}

\item{column}{index of the column which we are looking to merge nodes for}

\item{weights}{numeric vector containing weights for each row of \code{df}.}
}
\description{
Identifies the pair of nodes to merge which will minimize
the number of distinct links (ignoring weights).
}
