% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/worldmap.R
\name{GeographicRegionNames}
\alias{GeographicRegionNames}
\title{\code{GeographicRegionNames} Returns the list of unique geographic names that can be used when creating a WorldMap.}
\usage{
GeographicRegionNames(type)
}
\arguments{
\item{type...}{The name of the geographic region type. E.g.,\code{name}, \code{continent},}
}
\description{
\code{GeographicRegionNames} Returns the list of unique geographic names that can be used when creating a WorldMap.
}
\details{
Uses \code{\link{Zelig:zelimethods}}
which depend on the \code{\link{class}} of the first argument.
}

