% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/worldmap.R
\name{GeographicRegionRowNames}
\alias{GeographicRegionRowNames}
\title{\code{GeographicRegionRowNames} Names of geographic regions.
Returns the list of unique geographic names that
can be used when creating a WorldMap.}
\usage{
GeographicRegionRowNames(type)
}
\arguments{
\item{type}{The name of the geographic region type. See \code{\link{GeographicRegionTypes}}}
}
\description{
\code{GeographicRegionRowNames} Names of geographic regions.
Returns the list of unique geographic names that
can be used when creating a WorldMap.
}

