% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/worldmap.R
\name{GeographicRegionTypes}
\alias{GeographicRegionTypes}
\title{\code{GeographicRegionTypes} Types of Geographic Regions
The geographic region types that are available for refering
in a map. E.g., \code{name}, \code{continent},}
\usage{
GeographicRegionTypes()
}
\description{
\code{GeographicRegionTypes} Types of Geographic Regions
The geographic region types that are available for refering
in a map. E.g., \code{name}, \code{continent},
}

