% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/labeledscatterplot.R
\name{GetXlim}
\alias{GetXlim}
\title{ggplot xlim
\code{GetXlim}}
\usage{
GetXlim(p)
}
\arguments{
\item{p}{A \code{\link[ggplot2]{ggplot}} plot.}
}
\value{
A vector containing the minimum and and maximum value of x plotted.
}
\description{
ggplot xlim
\code{GetXlim}
}

