% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/scatterplot.R
\name{GetYlim}
\alias{GetYlim}
\title{ggplot ylim
\code{GetYlim}}
\usage{
GetYlim(p)
}
\arguments{
\item{p}{A \code{\link[ggplot2]{ggplot}} plot.}
}
\value{
A vector containing the minimum and and maximum value of y  plotted.
}
\description{
ggplot ylim
\code{GetYlim}
}

