% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/interactivelabeledscatterplot.R
\name{InteractiveLabeledScatterPlot}
\alias{InteractiveLabeledScatterPlot}
\title{\code{LabeledScatterPlot} Scatterplot with Labeled Points.}
\usage{
InteractiveLabeledScatterPlot(coords, group = NULL, row.labels = NULL,
  col.labels = NULL, group.name = "", space.substitute = "\\n",
  fixed.aspect = TRUE, colors = q.colors)
}
\arguments{
\item{coords}{The xy coordinates of the points.}

\item{group}{A factor indicating group membership for each point.}

\item{row.labels}{A vector of labels which will, if supplied, over-ride the rownames of coodinates.}

\item{col.labels}{A vector of labels which will, if supplied, over-ride the colnames of coodinates.}

\item{group.name}{Title for the legend (which only appears if group is not null).}

\item{fixed.aspect}{if true, forces the x and y dimensions to be on the same scale.}
}
\description{
\code{LabeledScatterPlot} Scatterplot with Labeled Points.
}

