% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/scatterplot.R
\name{IsOverlap}
\alias{IsOverlap}
\title{Box Overlaps
\code{IsOverlap} Tests if a box overlaps any other boxes.}
\usage{
IsOverlap(boxes, this.box)
}
\arguments{
\item{boxes}{A list of boxes, where each box is its coordinates.}

\item{this.box}{The box to check to see if it overlaps with the rest.}
}
\value{
TRUE if there is an overlap and false otherwise.
}
\description{
Box Overlaps
\code{IsOverlap} Tests if a box overlaps any other boxes.
}

