% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/scatterplot.R
\name{LabeledScatterPlot}
\alias{LabeledScatterPlot}
\alias{LabeledScatterPlot.default}
\title{Find better cootrdinates for text labels
\code{LabeledScatterPlot} Scatterplot with Labeled Points.}
\usage{
LabeledScatterPlot(coords, ...)

\method{LabeledScatterPlot}{default}(coords, group = NULL,
  row.labels = NULL, col.labels = NULL, title = "", legend.title = "",
  fixed.aspect = TRUE, auto.tidy = TRUE, colors = q.colors,
  auto.color = TRUE, general.color = "gray28", point.size = 2,
  label.size = 10, legend.size = 10, axis.title.size = 12,
  axis.label.size = 10, title.size = 12, overlap.fudge = 1,
  space.substitute = "\\n", ...)
}
\arguments{
\item{coords}{The xy coordinates of the points.}

\item{...}{Additional arguments.
#}

\item{group}{A factor indicating group membership for each point.}

\item{row.labels}{A vector of labels which will, if supplied, over-ride the rownames of coodinates.}

\item{col.labels}{A vector of labels which will, if supplied, over-ride the colnames of coodinates.}

\item{title}{Title for the plot.}

\item{legend.title}{Title for the legend (which only appears if group is not null).}

\item{fixed.aspect}{if true, forces the x and y dimensions to be on the same scale.}

\item{auto.tidy}{Move the labels around so that fewer overlap.}

\item{colors}{Colors that are cycled through where there is only one series, or, used to demarkate series where there are multiple series.}

\item{auto.color}{Automatically colors the points (if FALSE, the first color is used).}

\item{general.color}{The color to be used in axes and titles.}

\item{point.size}{Size of the dot representing the poing (i.e., the glyph).}

\item{label.size}{Font size of the text labels on the plot.}

\item{axis.title.size}{Font size of the axis titles.}

\item{axis.label.size}{Font size of the axis labels.}

\item{title.size}{Font size of the chart title.}

\item{overlap.fudge}{Determines the amount of space required between labels. A value of 1 corresponds to a best guess of
no overlap. The guess can be wrong, so the plot can be improved by modifying this value, which has a muliplier effect.}

\item{space.substitute}{Spaces in labels of points on plots are substituted with whatever is supplied.
By default, a return character is used(i.e., \link{"\n"}). To replace with a period, use \link{"\\."}.}

\item{object}{An object to be plotted.
#}

\item{row.description}{A title for the rows.
#}

\item{column.description}{A title for the columns.}
}
\value{
p A \code{\link[ggplot2]{ggplot}} plot.
}
\description{
Find better cootrdinates for text labels
\code{LabeledScatterPlot} Scatterplot with Labeled Points.
}
\section{Methods (by class)}{
\itemize{
\item \code{default}: Default labeled scatterplot
}}

