% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/namemaps.R
\name{makeNameMap}
\alias{makeNameMap}
\title{Make a list mapping between alternative names of countries}
\usage{
makeNameMap(column)
}
\arguments{
\item{column}{Usually either \code{"admin"} or \code{"name"}.}
}
\value{
A list of string vectors. The names are the primary country name and
  the vectors contain alternatives to those names.
}
\description{
Look at the admin-0 level data and ISOcodes package to find alternative names
of countries. The parameter \code{column} chooses the primary attribute (the
names of the list).
}

