% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sankeydiagram.R
\name{nodeDictionary}
\alias{nodeDictionary}
\title{nodeDictionary}
\usage{
nodeDictionary(list.of.factors)
}
\arguments{
\item{list.of.factors}{Factors representing variables in the data file.}
}
\description{
Creates a dictionary of the nodes.
}
