% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/outliersplot.R
\name{OutliersPlot}
\alias{OutliersPlot}
\alias{OutliersPlot.default}
\alias{OutliersPlot.lm}
\alias{OutliersPlot.smacof}
\alias{OutliersPlot.smacofR}
\title{Object Goodness-of-Fit Plot
\code{OutliersPlot} A generic function used to produce plots showing outliers in the data from
a model object.  The function invokes particular \code{\link{methods}}
which depend on the \code{\link{class}} of the first argument.}
\usage{
OutliersPlot(object)

\method{OutliersPlot}{default}(object, ...)

\method{OutliersPlot}{lm}(object, ...)

\method{OutliersPlot}{smacof}(object, ...)

\method{OutliersPlot}{smacofR}(object, ...)
}
\arguments{
\item{object}{An object for which a summary is desired..}

\item{...}{Additional arguments affecting the goodness-of-fit displayed.}
}
\description{
reports the goodness-of-fit of an object.
}
\section{Methods (by class)}{
\itemize{
\item \code{default}: Default outlier plot.

\item \code{lm}: Outliers plot for a linear model.

\item \code{smacof}: Outliers plot for a smacof object

\item \code{smacofR}: Outliers plot for a rectangular smacof object
}}
\examples{
# linear regression
x <- rnorm(10)
y <- rnorm(10) + x
mod <- lm(y ~ x)
OutliersPlot(mod)

# MDS - square
library(smacof)
data(breakfastDissimilarities)
mdsInterval <- smacofSym(breakfastDissimilarities[[4]],
    type = "interval", eps = 1e-12, itmax = 100000)
OutliersPlot(mdsInterval)
}

