% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/scatterplot.R
\name{OverlapIndex}
\alias{OverlapIndex}
\title{Box Overlaps
\code{OverlapIndex} Returns the index of the first overlapping box encountered.}
\usage{
OverlapIndex(boxes, this.box, this.box.index)
}
\arguments{
\item{boxes}{A list of boxes, where each box is its coordinates.}

\item{this.box}{The box to check to see if it overlaps with the rest.}

\item{this.box.index}{The index of this.box in the list of boxes.}
}
\value{
-1 if no overlaps found..
}
\description{
Box Overlaps
\code{OverlapIndex} Returns the index of the first overlapping box encountered.
}

