% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/qcolors.R
\docType{data}
\name{QColors}
\alias{QColors}
\title{\code{QColors} Colors used in plots in Displayr and Q.}
\format{An object of class \code{character} of length 14.}
\usage{
QColors
}
\description{
\code{QColors} Colors used in plots in Displayr and Q.
}
\keyword{datasets}

