% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/labeledscatterplot.R
\name{rectangle}
\alias{rectangle}
\title{\code{rectangle} The coordinates of a rectangle: bottom-left corner, width and height.}
\usage{
rectangle(x, y, w, h)
}
\arguments{
\item{x}{x coordinate of the middle of the box.}

\item{y}{y coordinate of the middle of the box.}

\item{w}{width of the box in the x-scale.}

\item{h}{height  of the box in the y-scale.}
}
\value{
A list containing the coordinates of a box, consisting
of the bottom left coordinates, width and height.
}
\description{
\code{rectangle} The coordinates of a rectangle: bottom-left corner, width and height.
}

