% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/labeledscatterplot.R
\name{rectangleOverlap}
\alias{rectangleOverlap}
\title{\code{rectangleOverlap} Checks to see if two recangles.}
\usage{
rectangleOverlap(a, b)
}
\arguments{
\item{a}{The first rectangle.}

\item{b}{The second recangle.}
}
\description{
\code{rectangleOverlap} Checks to see if two recangles.
}

