% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/scatterplot.R
\name{rectangleOverlapIndex}
\alias{rectangleOverlapIndex}
\title{\code{rectangleOverlapIndex} Returns the index of the first overlapping rectangle encountered.}
\usage{
rectangleOverlapIndex(rectangles, this.rectangle, this.rectangle.index)
}
\arguments{
\item{rectangles}{A list of rectangles, where each rectangle is its coordinates.}

\item{this.rectangle}{The box to check to see if it overlaps with the rest.}

\item{this.rectangle.index}{The index of this.rectangle in the list of boxes.}
}
\value{
-1 if no overlaps found..
}
\description{
\code{rectangleOverlapIndex} Returns the index of the first overlapping rectangle encountered.
}

