% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sankeydiagram.R
\name{SankeyDiagram}
\alias{SankeyDiagram}
\title{Sankey Diagram}
\usage{
SankeyDiagram(data, max.categories = 8)
}
\arguments{
\item{data}{A \code{\link{data.frame}} of variables.}

\item{max.categories}{When the number of unique values
of numeric data exceeds this value, the variable is quantized.}
}
\value{
A sankey diagram (HTMLwidget).
}
\description{
Creates a sankey diagram of the relationship between different variables.
}
\details{
Text variables are grouped as having text or not having text.
To see patterns with text variables, they should first be turned into
factors.
}

