% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scatterplotmatrix.R
\name{ScatterplotMatrix}
\alias{ScatterplotMatrix}
\title{Scatterplot matrix}
\usage{
ScatterplotMatrix(...)
}
\arguments{
\item{...}{The variables to plot. Can be data frames or vectors, but should
all be the same length as they will be bound together.}
}
\description{
Wrapper around the function \code{\link[psych]{pairs.panels}} from the
\code{psych} package. Ensures that the variable names come through nicely
from Displayr.
}

