% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/splinewithsimultaneousconfint.R
\name{SplineWithSimultaneousConfIntervals}
\alias{SplineWithSimultaneousConfIntervals}
\title{Plots spline with confidence intervals}
\usage{
SplineWithSimultaneousConfIntervals(
  outcome,
  predictor,
  type = "Binary Logit",
  subset = NULL,
  weights = NULL,
  seed = 42,
  number.draws = 30,
  confidence = 0.95
)
}
\arguments{
\item{outcome}{Outcome variable shown in the y-axis. This should be convertible using \code{AsNumeric}.}

\item{predictor}{Predictor variable shown in the x-axis. This should be a numeric or date variable.}

\item{type}{Regression link function used. Currently can be "Binary Logit" or "Linear".}

\item{subset}{Optional logical vector of the same length as \code{outcome}
and \code{predictor} specifying whether the observation should be included
in the analysis.}

\item{weights}{Optional numeric vector used as sampling weights.}

\item{seed}{Random seed for reproducibility}

\item{number.draws}{Number of possible trend lines to super-impose.}

\item{confidence}{Width of confidence interval shown around the spline.}
}
\description{
Plots spline with confidence intervals
}
