% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/labeledscatterplot.R
\name{SquaredDistance}
\alias{SquaredDistance}
\title{#
Squared distance between points
\code{SquaredDistance} Squared Euclidean distance between points}
\usage{
SquaredDistance(e1x, e1y, e2x, e2y)
}
\arguments{
\item{e1x}{The x coordinate of a point.}

\item{e1y}{The y coordinate of a point.}

\item{e2x}{The x coordinate of another point.}

\item{e2y}{The y coordinate of another point.}
}
\value{
The squared distance.
}
\description{
#
Squared distance between points
\code{SquaredDistance} Squared Euclidean distance between points
}

