% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/worldmap.R
\name{WorldChoropleth}
\alias{WorldChoropleth}
\title{World map based on \code{choroplethr}}
\usage{
WorldChoropleth(..., remove.regions = "antarctica")
}
\description{
Similar to \code{\link{WorldMap}} but does the plotting using
\code{choroplethr} rather than \code{leaflet}.
}

