% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/worldmap.R
\name{WorldMap}
\alias{WorldMap}
\title{\code{Linear Regression} Linear Regression.}
\usage{
WorldMap(table, treat.zeros.differently = TRUE, remove.last.column = TRUE,
  remove.last.row = TRUE, type = "name", colors = c("#CCF3FF", "#23B0DB"),
  legend.title = "", only.show.regions.in.table = FALSE,
  remove.antarctica = TRUE, add.detail = FALSE)
}
\arguments{
\item{object}{An object for which a summary is desired..}

\item{...}{Additional arguments affecting the goodness-of-fit displayed.}
}
\description{
reports the goodness-of-fit of an object.
}
\details{
Uses \code{\link{Zelig:zelimethods}}
which depend on the \code{\link{class}} of the first argument.
}

