% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/labeledscatterplot.R
\name{BoxCoordinates}
\alias{BoxCoordinates}
\title{Box coordinates
\code{BoxOverlap} Computes the positions of the corner of a box.}
\usage{
BoxCoordinates(x, y, w, h)
}
\arguments{
\item{x}{x coordinate of the middle of the box.}

\item{y}{y coordinate of the middle of the box.}

\item{w}{width of the box in the x-scale.}

\item{h}{height  of the box in the y-scale.}
}
\value{
A vector containing the coordinates of a box.
}
\description{
Box coordinates
\code{BoxOverlap} Computes the positions of the corner of a box.
}

