% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sankeydiagram.R
\name{categorizeData}
\alias{categorizeData}
\title{categorizeData}
\usage{
categorizeData(data, max.categories, share.values)
}
\arguments{
\item{data}{A \code{\link{data.frame}} or \code{\link{list}} of variables.}

\item{max.categories}{When the number of unique values}

\item{share.values}{Whether the values in each variable are expected to be the same.
of numeric data exceeds this value, the variable is quantized.}
}
\description{
Creates a dictionary of the nodes.
Quantizes numeric variables.
}
