% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sankeydiagram.R
\name{categorizeVariable}
\alias{categorizeVariable}
\title{categorizeVariable}
\usage{
categorizeVariable(x, max.categories, var.name)
}
\arguments{
\item{x}{A variable.}

\item{max.categories}{With numeric
variables, \code{\link{cut}} is used, unless \code{max.categories} has a value of
2 and there are missing values, in which case the data is turned into a factor
with levels of "Data" and "NA"). With factors, the two smallest
categories are merged (based on unweighted data). With ordered factors, the
smallest category is merged with the smallest adjacent category.
Missing values are not merged. With character variables, the data is merged into missing
versus non-missing data.}

\item{var.name}{The name of the variable.
of numeric data exceeds this value, the variable is quantized.}
}
\description{
Quantizes a numeric variable.
}

