% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/worldmap.R
\name{WorldChoropleth}
\alias{Choropleth}
\alias{WorldChoropleth}
\title{World map based on \code{choroplethr}}
\usage{
WorldChoropleth(..., remove.regions = "antarctica")

Choropleth(df, region, treat.NA.as.0 = FALSE,
  only.show.regions.in.table = FALSE, remove.last.row = FALSE,
  legend.title = "", remove.regions = NULL)
}
\arguments{
\item{...}{Other parameters to pass to \code{Choropleth}.}

\item{remove.regions}{The regions to remove, even if they are in the table.}

\item{df}{A matrix, two-dimensional array, table or vector, containing the
data to be plotted. The \code{\link{rownames}} (or \code{\link{names}} in
the case of a vector) should contain the names of the geographic entities
to be plotted. If it is a matrix with more than 1 column, the first will be
used as the values on the map.}

\item{region}{The region of the world to map. This can either be
\dQuote{world} or the name of a country.}

\item{treat.NA.as.0}{Plots any \code{NA} values in the data and any
geographical entities without data with as having a 0 value.}

\item{only.show.regions.in.table}{When TRUE, only geographic entities that
are included in the table are shown on the table.}

\item{remove.last.row}{Deletes the bottom row of the table prior to creating
the map, unless the table is a vector or only has one row.}

\item{legend.title}{The text to appear above the legend.}
}
\description{
Similar to \code{\link{WorldMap}} but does the plotting using
\code{choroplethr} rather than \code{leaflet}.

Base function for plotting a map
}
\section{Functions}{
\itemize{
\item \code{WorldChoropleth}: A choropleth of the countries of the world.
}}

