% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sankeydiagram.R
\name{computeLinks}
\alias{computeLinks}
\title{computeLinks}
\usage{
computeLinks(data)
}
\arguments{
\item{data}{A \code{\link{data.frame}} or \code{\link{list}} of variables.}
}
\description{
Computes the links between the nodes, so that can be expressed as a network.
}
