% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{map.coordinates}
\alias{admin1.coordinates}
\alias{map.coordinates}
\title{Admin-0 level coordinates for the world}
\format{A SpatialPolygonsDataFrame with data frame with 177 rows and 63 columns:}
\source{
\url{http://www.naturalearthdata.com/downloads/50m-cultural-vectors/}

\url{http://www.naturalearthdata.com/downloads/10m-cultural-vectors/}
}
\usage{
map.coordinates

admin1.coordinates
}
\description{
Admin-0 level coordinates for the world

Admin-1 level coordinates for the world
}
\keyword{datasets}

