% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sankeydiagram.R
\name{findNodesToMerge}
\alias{findNodesToMerge}
\title{findNodesToMerge}
\usage{
findNodesToMerge(df, column, weights = NULL)
}
\arguments{
\item{df}{dataframe containing factor variables only.}

\item{column}{index of the column which we are looking to merge nodes for}

\item{weights}{numeric vector containing weights for each row of \code{df}.}
}
\description{
Identifies the pair of nodes to merge which will minimize
distinct links. Note this uses the weights, so adding multiple
small links may be preferred to adding a single heavy link.
Note also that the link to the merged node i.e. (A,B+C) is the
has a weight that is the sum of (A,B) and (A,C). The relative
proportions are not taken into account.
}
