% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/interactivelabeledscatterplot.R
\name{InteractiveLabeledScatterPlot}
\alias{InteractiveLabeledScatterPlot}
\title{\code{LabeledScatterPlot} Scatterplot with Labeled Points.}
\usage{
InteractiveLabeledScatterPlot(coords, group = NULL, row.labels = NULL,
  column.labels = NULL, group.name = "", tooltip.text = NULL,
  fixed.aspect = FALSE, colors = q.colors, auto.color = TRUE,
  legend.width = 150)
}
\arguments{
\item{coords}{The xy coordinates of the points.}

\item{group}{A factor indicating group membership for each point.}

\item{row.labels}{A vector of labels which will, if supplied, over-ride the rownames of coodinates.}

\item{column.labels}{A vector of labels which will, if supplied, over-ride the colnames of coodinates.}

\item{group.name}{The title to appear above the legend, which indicates group membership.}

\item{tooltip.text}{The text to appear on tooltips.}

\item{fixed.aspect}{If true, forces the x and y dimensions to be on the same scale.}

\item{colors}{Colors that are cycled through where there is only one series, or, used to demarkate series where there are multiple series.}

\item{auto.color}{Automatically colors the points (if FALSE, the first color is used).}

\item{legend.width}{Legend area width, in pixels.}

\item{group.name}{Title for the legend (which only appears if group is not null).}
}
\description{
\code{LabeledScatterPlot} Scatterplot with Labeled Points.
}

