% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/labeledscatterplot.R
\name{LabeledScatterPlot}
\alias{LabeledScatterPlot}
\alias{LabeledScatterPlot.default}
\alias{LabeledScatterPlot.smacof}
\alias{LabeledScatterPlot.smacofB}
\alias{LabeledScatterPlot.smacofR}
\title{Scatterplot with Labeled Points.}
\usage{
LabeledScatterPlot(object, ...)

\method{LabeledScatterPlot}{default}(object, group = NULL,
  row.labels = NULL, column.labels = NULL, title = "", group.name = "",
  fixed.aspect = FALSE, colors = QColors, auto.color = TRUE,
  general.color = "gray28", point.size = 1, label.font.size = 10,
  legend.font.size = 10, axis.title.font.size = 12,
  axis.label.font.size = 10, title.font.size = 12)

\method{LabeledScatterPlot}{smacof}(object, ...)

\method{LabeledScatterPlot}{smacofB}(object, ...)

\method{LabeledScatterPlot}{smacofR}(object, row.description = NULL,
  column.description = NULL, ...)
}
\arguments{
\item{object}{The xy coordinates of the points.}

\item{...}{Other plotting arguments.}

\item{group}{A factor indicating group membership for each point.}

\item{row.labels}{A vector of labels which will, if supplied, over-ride the rownames of coodinates.}

\item{column.labels}{A vector of labels which will, if supplied, over-ride the colnames of coodinates.}

\item{title}{Title for the plot.}

\item{group.name}{Title for the legend (which only appears if group is not null).}

\item{fixed.aspect}{if true, forces the x and y dimensions to be on the same scale.}

\item{colors}{Colors that are cycled through where there is only one series, or, used to demarkate series where there are multiple series.}

\item{auto.color}{Automatically colors the points (if FALSE, the first color is used).}

\item{general.color}{The color to be used in axes and titles.}

\item{point.size}{Size of the dot representing the poing (i.e., the glyph).}

\item{label.font.size}{Font size of the text labels on the plot.}

\item{legend.font.size}{Font size of the for the plot.}

\item{axis.title.font.size}{Font size of the axis titles.}

\item{axis.label.font.size}{Font size of the axis labels.}

\item{title.font.size}{Font size of the chart title.}
}
\value{
A \code{\link[ggplot2]{ggplot}} plot.
}
\description{
\code{LabeledScatterPlot} Scatterplot with Labeled Points.

\code{LabeledScatterPlot.default} Scatterplot with Labeled Points.
}

