% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/scatterplot.R
\name{LabeledScatterPlot}
\alias{LabeledScatterPlot}
\alias{LabeledScatterPlot.default}
\title{Find better cootrdinates for text labels
\code{LabeledScatterPlot} Scatterplot with Labeled Points.}
\usage{
LabeledScatterPlot(coords, ...)

\method{LabeledScatterPlot}{default}(coords, group = NULL,
  row.labels = NULL, col.labels = NULL, title = "", legend.title = "",
  fixed.aspect = TRUE, auto.tidy = TRUE, colors = qColors,
  auto.color = 10, general.color = "gray28", cex = 1, ...)
}
\arguments{
\item{coords}{The xy coordinates of the points.}

\item{...}{Additional arguments.
#}

\item{group}{A factor indicating group membership for each point.}

\item{row.labels}{A vector of labels which will, if supplied, over-ride the rownames of coodinates.}

\item{col.labels}{A vector of labels which will, if supplied, over-ride the colnames of coodinates.}

\item{title}{Title for the plot.}

\item{legend.title}{Title for the legend (which only appears if group is not null).}

\item{fixed.aspect}{if true, forces the x and y dimensions to be on the same scale.}

\item{auto.tidy}{Move the labels around so that fewer overlap.}

\item{colors}{Colors that are cycled through where there is only one series, or, used to demarkate series where there are multiple series.}

\item{auto.color}{When the number of rows of coords is less than or equal to this, a single color is used to label the points.  Otherwise, they cycle throuogh the colors.}

\item{general.color}{The color to be used in axes and titles.}

\item{cex}{Relative font, line and glyph size.}

\item{coords}{Relative font, line and glyph size.}

\item{object}{An object to be plotted.
#}

\item{row.description}{A title for the rows.
#}

\item{column.description}{A title for the columns.}
}
\value{
p A \code{\link[ggplot2]{ggplot}} plot.
}
\description{
Find better cootrdinates for text labels
\code{LabeledScatterPlot} Scatterplot with Labeled Points.
}
\section{Methods (by class)}{
\itemize{
\item \code{default}: Default labeled scatterplot
}}

