% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sankeydiagram.R
\name{mergeNodes}
\alias{mergeNodes}
\title{Merge levels in a factor}
\usage{
mergeNodes(x, old.nodes)
}
\arguments{
\item{x}{factor variable}

\item{old.nodes}{a vector containing factor levels to merge}
}
\description{
Merge levels in a factor
}
