% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/qcolors.R
\docType{data}
\name{QColors}
\alias{QColors}
\title{\code{QColors} Colors used in plots in Displayr and Q.}
\format{\preformatted{ chr [1:14] "#5B9BD5FF" "#ED7D31FF" "#A5A5A5FF" "#1EC000FF" ...
}}
\usage{
QColors
}
\description{
\code{QColors} Colors used in plots in Displayr and Q.
}
\keyword{datasets}

