% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/labeledscatterplot.R
\name{ReducePointAndLabelOverlap}
\alias{ReducePointAndLabelOverlap}
\title{Find better cootrdinates for text labels
\code{GetYlim} Identifies a better set of coordinates to place labels in a labeled ggplot scatterplot.}
\usage{
ReducePointAndLabelOverlap(p, label.size, do.nothing, fixed.aspect, tstep,
  rstep, overlap.fudge)
}
\arguments{
\item{p}{A \code{\link[ggplot2]{ggplot}} plot.}

\item{do.nothing}{If TRUE, the algorithm does nothing, other than return values.}

\item{fixed.aspect}{If TRUE, forces the x and y dimensions to be on the same scale.}

\item{tstep}{The angle (theta) step size as the algorithm spirals out.}

\item{rstep}{The radius step size (in standard deviations) as the algorithm spirals out.}

\item{overlap.fudge}{Determines the amount of space required between labels. A value of 1 corresponds to a best guess of
no overlap. The guess can be wrong, so the plot can be improved by modifying this value, which has a muliplier effect.}

\item{plot.width}{The width of the plotting area (in inches).}

\item{plot.height}{The height of the plotting area (in inches).}
}
\value{
dimensions Width and height of the text to be plotted in terms of the scale of x and y.
}
\description{
Find better cootrdinates for text labels
\code{GetYlim} Identifies a better set of coordinates to place labels in a labeled ggplot scatterplot.
}

