% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sankeydiagram.R
\name{SankeyDiagram}
\alias{SankeyDiagram}
\title{Sankey Diagram}
\usage{
SankeyDiagram(data, max.categories = 8, subset = NULL, weights = NULL,
  font.size = 12, font.family = "Times New Roman", colors = NULL,
  link.color = c("None", "Source", "Target")[1], node.width = 30)
}
\arguments{
\item{data}{A \code{\link{data.frame}} of variables.}

\item{max.categories}{When the number of unique values
of numeric data exceeds this value, the variable is quantized.}

\item{subset}{An optional vector specifying a subset of observations to be
used in the fitting process.}

\item{weights}{An optional vector of sampling weights.}

\item{font.size}{Font size of node labels.}

\item{font.family}{Font family of node labels.}

\item{colors}{Colors of the nodes, supplied as a vector of hex colors.}

\item{link.color}{One of \code{"None", "Source", "Target"}. This specifies how whether
the links are shown in grey (None), the same color as the source node, or the same
color as the target node.}

\item{node.width}{Width of the width.}
}
\value{
A sankey diagram (HTMLwidget).
}
\description{
Creates a sankey diagram of the relationship between different variables.
}
\details{
Text variables are grouped as having text or not having text.
To see patterns with text variables, they should first be turned into
factors.
}
