% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sankeydiagram.R
\name{SankeyDiagram}
\alias{SankeyDiagram}
\title{Sankey Diagram}
\usage{
SankeyDiagram(data = NULL, links.and.nodes = NULL,
  output.data.only = FALSE, max.categories = 8, subset = NULL,
  weights = NULL, font.size = 12, font.family = "Times New Roman",
  font.unit = "px", colors = NULL, link.color = c("None", "Source",
  "Target", "First variable", "Last variable")[1],
  variables.share.values = FALSE, label.show.varname = TRUE,
  label.max.length = 100, label.show.counts = FALSE,
  label.show.percentages = FALSE, node.width = 30, node.padding = 10,
  hovertext.show.percentages = FALSE)
}
\arguments{
\item{data}{A \code{\link{data.frame}} of variables.}

\item{links.and.nodes}{Input a list structure from a previous call of SankeyDiagram
with \code{output.data.only = TRUE}.}

\item{output.data.only}{Only outputs the link and node structure without printing the sankey diagram.}

\item{max.categories}{When the number of unique values
of numeric data exceeds this value, the variable is quantized.}

\item{subset}{An optional vector specifying a subset of observations to be
used in the fitting process.}

\item{weights}{An optional vector of weights. If not provided, the width of
the links is proportional to the number of rows in the data frame that correspond
to the path (flow).}

\item{font.size}{Font size of node labels.}

\item{font.family}{Font family of node labels.}

\item{font.unit}{One of "px" of "pt". By default all font sizes are specified in terms of
pixels ("px"). But changing this to "pt" will mean that the font sizes will be in terms
points ("pt"), which will be consistent with font sizes in text boxes.}

\item{colors}{Colors of the nodes, supplied as a vector of hex colors.
Transparency values (alpha) will be ignored.}

\item{link.color}{One of \code{"None", "Source", "Target", "First variable",
"Last variable"}. This specifies whether the links are shown in grey (None);
the same color as the source node; the same color as the target node; or
the same color as node in the first or last variable.}

\item{variables.share.values}{If \code{TRUE}, and \code{link.color = "Source"}
or \code{"Target"}, then the same set colors will be used for each variable.}

\item{label.show.varname}{Show variable name in node label.}

\item{label.max.length}{Maximum number of characters in each node label.}

\item{label.show.counts}{Append count data to node labels.}

\item{label.show.percentages}{Append percentages to node labels.}

\item{node.width}{Width of the width.}

\item{node.padding}{Vertical space between nodes.}

\item{hovertext.show.percentages}{Show percentages instead of counts in the hovertext.}
}
\value{
A sankey diagram (HTMLwidget).
}
\description{
Creates a sankey diagram of the relationship between different variables.
}
\details{
Text variables are grouped as having text or not having text.
To see patterns with text variables, they should first be turned into
factors.
}
