% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/splinewithsimultaneousconfint.R
\name{SplineWithSimultaneousConfIntervals}
\alias{SplineWithSimultaneousConfIntervals}
\title{Plots spline with confidence intervals}
\usage{
SplineWithSimultaneousConfIntervals(
  outcome,
  predictor,
  type = "Binary Logit",
  subset = NULL,
  weights = NULL,
  seed = 42,
  number.draws = 30,
  confidence = 0.95,
  mean.color = "#000000",
  mean.weight = 2,
  draw.color = "#33333366",
  draw.weight = 1,
  ci.color = "#FF000033",
  y.tick.format = NULL,
  y.hovertext.format = ".1f",
  title = NULL,
  subtitle = NULL,
  x.title = NULL,
  y.title = NULL,
  global.font.family = "Open Sans",
  global.font.color = rgb(44, 44, 44, maxColorValue = 255),
  title.font.size = 12,
  subtitle.font.size = 8,
  hovertext.font.size = 9,
  y.title.font.size = 10,
  x.title.font.size = 10,
  y.tick.font.size = 9,
  x.tick.font.size = 9,
  font.units = "pt",
  trim.padding = FALSE,
  ...
)
}
\arguments{
\item{outcome}{Outcome variable shown in the y-axis. This should be convertible using \code{AsNumeric}.}

\item{predictor}{Predictor variable shown in the x-axis. This should be a numeric or date variable.}

\item{type}{Regression link function used. Currently can be "Binary Logit" or "Linear".}

\item{subset}{Optional logical vector of the same length as \code{outcome}
and \code{predictor} specifying whether the observation should be included
in the analysis.}

\item{weights}{Optional numeric vector used as sampling weights.}

\item{seed}{Random seed for reproducibility}

\item{number.draws}{Number of possible trend lines to super-impose.}

\item{confidence}{Width of confidence interval shown around the spline.}

\item{mean.color}{Line color of the mean (or predicted) line.}

\item{mean.weight}{Line weight of the mean line.}

\item{draw.color}{Color of the possible trend lines.}

\item{draw.weight}{Line weight of the possible trend line.}

\item{ci.color}{Color of the ribbon showing the confidence interval.}

\item{y.tick.format}{A string representing a d3 formatting code.
See https://github.com/d3/d3/blob/master/API.md#number-formats-d3-format}

\item{y.hovertext.format}{A string representing a d3 formatting code
See https://github.com/d3/d3/blob/master/API.md#number-formats-d3-format}

\item{title}{Character; chart title.}

\item{subtitle}{Character}

\item{x.title}{Character, x-axis title; defaults to chart input values;
to turn off set to "FALSE".}

\item{y.title}{Character, y-axis title; defaults to chart input values;
to turn off set to "FALSE".}

\item{global.font.family}{Character; font family for all occurrences of any
font attribute for the chart unless specified individually.}

\item{global.font.color}{Global font color as a named color in character format
(e.g. "black") or an a hex code.}

\item{title.font.size}{Integer; Title font size; default = 10.}

\item{subtitle.font.size}{Integer; subtitle font size}

\item{hovertext.font.size}{Font size of hover text.}

\item{y.title.font.size}{Integer; y-axis title font size}

\item{x.title.font.size}{Integer; x-axis title font size}

\item{y.tick.font.size}{Integer; y-axis tick label font size}

\item{x.tick.font.size}{Integer; x-axis tick label font size}

\item{font.units}{One of "px" of "pt". By default all font sizes are specified in terms of
pixels ("px"). But changing this to "pt" will mean that the font sizes will be in terms
points ("pt"), which will be consistent with font sizes in text boxes.}

\item{trim.padding}{(Deprecated)Logical; whether to remove padding around plotly chart.
Default is set to false so that output is the same as old charts.}

\item{...}{Other parameters to pass to flipStandardCharts::Line.}
}
\description{
Plots spline with confidence intervals
}
